+++ src/sigproc.h
@@ -506,14 +506,14 @@ public:
     std::vector<FFTTYPE> in;
     std::vector<std::complex<FFTTYPE> > out;
 #ifdef FFT_FFTW3
-    inline void setInput(int n, FFTTYPE value){m_fftw3_in[n] = value;}
-    inline std::complex<FFTTYPE> getOutput(int n){return make_complex(m_fftw3_out[n]);}
+    inline void setInput(size_t n, FFTTYPE value){m_fftw3_in[n] = value;}
+    inline std::complex<FFTTYPE> getOutput(size_t n){return make_complex(m_fftw3_out[n]);}
     inline std::complex<FFTTYPE> getDCOutput(){return make_complex(m_fftw3_out[0]);} // Avoid index checking
-    inline std::complex<FFTTYPE> getMidOutput(int n){return make_complex(m_fftw3_out[n]);} // Avoid index checking
+    inline std::complex<FFTTYPE> getMidOutput(size_t n){return make_complex(m_fftw3_out[n]);} // Avoid index checking
     inline std::complex<FFTTYPE> getNyquistOutput(){return make_complex(m_fftw3_out[m_size/2]);}// Avoid index checking
 #elif FFT_FFTREAL
-    inline void setInput(int n, FFTTYPE value){in[n] = value;}
-    inline std::complex<FFTTYPE> getOutput(int n){
+    inline void setInput(size_t n, FFTTYPE value){in[n] = value;}
+    inline std::complex<FFTTYPE> getOutput(size_t n){
         if(n==0)
             return make_complex(m_fftreal_out[0], 0.0);
         if(n==m_size/2)
@@ -521,7 +521,7 @@ public:
         return make_complex(m_fftreal_out[n], -m_fftreal_out[m_size/2+n]);
     }
     inline std::complex<FFTTYPE> getDCOutput(){return make_complex(m_fftreal_out[0], 0.0);} // Avoid index checking
-    inline std::complex<FFTTYPE> getMidOutput(int n){return make_complex(m_fftreal_out[n], -m_fftreal_out[m_size/2+n]);} // Avoid index checking
+    inline std::complex<FFTTYPE> getMidOutput(size_t n){return make_complex(m_fftreal_out[n], -m_fftreal_out[m_size/2+n]);} // Avoid index checking
     inline std::complex<FFTTYPE> getNyquistOutput(){return make_complex(m_fftreal_out[m_size/2],0.0);}// Avoid index checking
 #endif
